+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSETWJ8Z4BFMGH2TFK5RE32M"
+        "01KSF0AB23E412YGAPA4AYBDJP"
     )
     r"""The unique identifier of the data source"""