+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSETWJ95EK4HFA2TXYREEMXR"
+        "01KSF0AB2723CYQBBXDRSDH2YZ"
     )
     r"""The unique identifier of the data source"""